refactor: 입출력 뷰 분리
