+++ src/main/java/lotto/view/InputView.java
@@ -1,0 +1,9 @@
+package lotto.view;
+
+import camp.nextstep.edu.missionutils.Console;
+
+public class InputView {
+    public static String read() {
+        return Console.readLine();
+    }
+}
